+++ packages/webpack-cli/lib/utils/__tests__/prompt-installation.test.js
@@ -8,20 +8,16 @@ jest.setMock('enquirer', {
     prompt: jest.fn(),
 });
 
-jest.setMock('../run-command', {
-    runCommand: jest.fn(),
-});
+jest.setMock('../run-command', jest.fn());
 
-jest.setMock('../package-exists', {
-    packageExists: jest.fn(),
-});
+jest.setMock('../package-exists', jest.fn());
 
 jest.setMock('../get-package-manager', jest.fn());
 
 const getPackageManager = require('../get-package-manager');
-const { packageExists } = require('../package-exists');
-const { promptInstallation } = require('../prompt-installation');
-const { runCommand } = require('../run-command');
+const packageExists = require('../package-exists');
+const promptInstallation = require('../prompt-installation');
+const runCommand = require('../run-command');
 const { prompt } = require('enquirer');
 
 describe('promptInstallation', () => {
