+++ packages/webpack-cli/lib/utils/package-exists.js
@@ -7,6 +7,4 @@ function packageExists(packageName) {
     }
 }
 
-module.exports = {
-    packageExists,
-};
+module.exports = packageExists;
