+++ packages/webpack-cli/lib/utils/to-kebab-case.js
@@ -7,4 +7,4 @@ function toKebabCase(str) {
     return str.replace(/([a-z0-9])([A-Z])/g, '$1-$2').toLowerCase();
 }
 
-module.exports = { toKebabCase };
+module.exports = toKebabCase;
